Google Chrome Speech Recognition
Added support for writting e-mails using speech recognition over Google
Chrome.
In the NEW E-MAIL page, the user can see a microphone icon. When the
user clicks the icon, the application begins to listen up to the user
stops the recording.
The recognized speech will be written automatically on the text box.